Refactored unique words to utility
